# src/main/java/br/com/dnafutsal/backend/identity/api/RefreshTokenRequest.java
package br.com.dnafutsal.backend.identity.api;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;

public record RefreshTokenRequest(@NotBlank @Size(max = 200) String refreshToken) {
}